Add new file “CardsAdminViewModel”
